Migrate role's listing, creation and edition
