+++ angular/src/app/roles/create-role/create-role-dialog.component.ts
@@ -1,0 +1,105 @@
+import { Component, Injector, OnInit } from '@angular/core';
+import {
+  MatDialogRef,
+  MatCheckboxChange
+} from '@angular/material';
+import * as _ from 'lodash';
+import {
+  RoleServiceProxy,
+  RoleDto,
+  ListResultDtoOfPermissionDto,
+  PermissionDto,
+  CreateRoleDto
+} from '@shared/service-proxies/service-proxies';
+import { AppComponentBase } from '@shared/app-component-base';
+import { finalize } from 'rxjs/operators';
+
+@Component({
+  templateUrl: 'create-role-dialog.component.html',
+  styles: [
+    `
+      mat-form-field {
+        width: 100%;
+      }
+      mat-checkbox {
+        padding-bottom: 5px;
+      }
+    `
+  ]
+})
+export class CreateRoleDialogComponent extends AppComponentBase
+  implements OnInit {
+  role: RoleDto = new RoleDto();
+  permissions: PermissionDto[] = [];
+  grantedPermissionNames: string[] = [];
+  checkedPermissionsMap: { [key: string]: boolean } = {};
+  defaultPermissionCheckedStatus = true;
+  saving = false;
+
+  constructor(
+    injector: Injector,
+    private _roleService: RoleServiceProxy,
+    private _dialogRef: MatDialogRef<CreateRoleDialogComponent>
+  ) {
+    super(injector);
+  }
+
+  ngOnInit(): void {
+    this._roleService
+      .getAllPermissions()
+      .subscribe((result: ListResultDtoOfPermissionDto) => {
+        this.permissions = result.items;
+        this.setInitialPermissionsStatus();
+      });
+  }
+
+  setInitialPermissionsStatus(): void {
+    _.map(this.permissions, item => {
+      this.checkedPermissionsMap[item.name] = this.defaultPermissionCheckedStatus
+    });
+  }
+
+  isPermissionChecked(permissionName: string): boolean {
+    // just return default permission checked status
+    // it's better to use a setting
+    return this.defaultPermissionCheckedStatus;
+  }
+
+  onPermissionChange(permission: PermissionDto, $event: MatCheckboxChange) {
+    this.checkedPermissionsMap[permission.name] = $event.checked;
+  }
+
+  getCheckedPermissions(): string[] {
+    const permissions: string[] = [];
+    _.forEach(this.checkedPermissionsMap, function (value, key) {
+      if (value) {
+        permissions.push(key);
+      }
+    });
+    return permissions;
+  }
+
+  save(): void {
+    this.saving = true;
+
+    const role = new RoleDto();
+    role.permissions = this.getCheckedPermissions();
+    role.init(this.role);
+
+    const createRole = new CreateRoleDto();
+    createRole.init(role);
+    this._roleService.create(createRole)
+      .pipe(
+        finalize(() => {
+          this.saving = false;
+        })
+      ).subscribe(() => {
+        this.notify.info(this.l('SavedSuccessfully'));
+        this.close(true);
+      });
+  }
+
+  close(result: any): void {
+    this._dialogRef.close(result);
+  }
+}
